bard microservice

+++ thresh/migrations/2022-06-22-221601_create_matches/up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE matches (
   match_id VARCHAR NOT NULL PRIMARY KEY,
-  game_datetime INTEGER NOT NULL,
-  game_length INTEGER NOT NULL,
+  game_datetime BIGINT NOT NULL,
+  game_length INT NOT NULL,
   tft_set_core_name VARCHAR NOT NULL,
   region VARCHAR
 )
